Models rspec PR

+++ db/migrate/20230213132457_create_food.rb
@@ -3,8 +3,8 @@ class CreateFood < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :measurement_unit
-      t.decimal :price
-      t.decimal :quantity
+      t.float :price
+      t.float :quantity
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
